+++ src/app/modules/browse-records-page/browse-records-page.component.ts
@@ -19,6 +19,7 @@ import Resource = fhir.Resource;
 import { ResourceTableComponent } from '../resource-table/resource-table.component';
 import { SelectRecordsService } from '../../shared/select-records/select-records.service';
 import { Sort } from '@angular/material/sort';
+import { getPluralFormOfRecordName } from '../../shared/utils';
 
 /**
  * Component for browsing public data (ResearchStudies and Variables).
@@ -47,6 +48,12 @@ export class BrowseRecordsPageComponent
   sort: { [resourceType: string]: Sort } = {
     ResearchStudy: {
       active: 'title',
+      // MatTable shows sort order icons in reverse (see comment to PR on LF-1905).
+      direction: 'desc'
+    },
+    Variable: {
+      active: 'display_name',
+      // MatTable shows sort order icons in reverse (see comment to PR on LF-1905).
       direction: 'desc'
     }
   };
@@ -82,6 +89,11 @@ export class BrowseRecordsPageComponent
         })
     );
   }
+
+  /**
+   * Returns plural form of resource type name.
+   */
+  getPluralFormOfRecordName = getPluralFormOfRecordName;
 
   ngOnInit(): void {}
 
@@ -122,26 +134,6 @@ export class BrowseRecordsPageComponent
 
       const resourceType = this.visibleResourceTypes[0];
       this.loadFirstPage(resourceType);
-    });
-  }
-
-  /**
-   * Returns text for the remove tab button.
-   */
-  getRemoveTabButtonText(resourceType: string): string {
-    return `Remove ${this.getPluralFormOfResourceType(resourceType)} tab`;
-  }
-
-  /**
-   * Returns plural form of resource type name.
-   */
-  getPluralFormOfResourceType(resourceType: string): string {
-    const tabName = this.resourceType2TabName[resourceType] || resourceType;
-    return tabName.replace(/(.*)(.)/, (_, $1, $2) => {
-      if ($2 === 'y') {
-        return $1 + 'ies';
-      }
-      return _ + 's';
     });
   }
 
@@ -220,6 +212,7 @@ export class BrowseRecordsPageComponent
     if (!sort) {
       return '';
     }
+    // MatTable shows sort order icons in reverse (see comment to PR on LF-1905).
     return `_sort=${sort.direction === 'asc' ? '-' : ''}${sort.active}`;
   }