+++ src/app/modules/search-parameter-group/search-parameter-group.component.ts
@@ -46,6 +46,7 @@ export class SearchParameterGroupComponent
   extends BaseControlValueAccessor<SearchParameterGroup>
   implements OnInit {
   @Input() inputResourceType = '';
+  @Input() allowAddRemoveButtons = true;
   @ViewChild('resourceTypeInput') resourceTypeInput: ElementRef;
   @ViewChildren(SearchParameterComponent)
   searchParameterComponents: QueryList<SearchParameterComponent>;