+++ CHANGELOG.md
@@ -3,10 +3,26 @@
 This log documents the significant changes for each release.
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [5.0.3] - 2022-08-23
+## [5.0.6] - 2022-08-26
 ### Changed
 - Replace mat-autocomplete with autocomplete-lhc controls.
 - Match list value on search parameter value autocomplete controls.
+
+## [5.0.5] - 2022-08-25
+### Added
+- Sorting to the variable table (for the alpha version).
+
+## [5.0.4] - 2022-08-25
+### Added
+- Simple prototype of the component for records in cart for the alpha version.
+- The ability to switch to the alpha version using the menu on the version
+  number link.
+### Fixed
+- Incorrect query parameter _type replaced with _format.
+
+## [5.0.3] - 2022-08-24
+### Changed
+- Removed prev folder.
 
 ## [5.0.2] - 2022-07-26
 ### Fixed